requirements.txt
Flask==3.0.0
openai>=2.0.0
PyPDF2==3.0.1
python-docx==1.1.0
cryptography==42.0.5
werkzeug==3.0.1
python-dotenv==1.0.1
